Move SpanLineEnumerator to System.Text
